refactor: modularize diagram toolbar
